fix no leak tests for RepeatedTask (#569)
Motivation:

The tests that test that we don't leak RepeatedTasks/EventLoops were a
bit racy so this needed fixing.

Modifications:

fixed the races in the tests

Result:

fixes #565
